refactor: group sidepanel and its components
