+++ frontend/src/components/Sidepanel/RoomInfo/index.tsx
@@ -10,11 +10,11 @@ import {
 } from "lucide-react";
 import { AnimatePresence, motion } from "motion/react";
 import { useState } from "react";
-import { useRoomSelection } from "../hooks/useRoomSelection";
-import { editRoom } from "../services/roomsService";
-import { EXPAND_COLLAPSE_TRANSITION } from "../utils/motionTransitions";
-import { renderValue } from "../utils/renderValue";
-import RoomModal from "./RoomModal";
+import { useRoomSelection } from "../../../hooks/useRoomSelection";
+import { editRoom } from "../../../services/roomsService";
+import { EXPAND_COLLAPSE_TRANSITION } from "../../../utils/motionTransitions";
+import { renderValue } from "../../../utils/renderValue";
+import RoomModal from "../../RoomModal";
 import "./SidePanel.css";
 
 type RoomInfoProps = {
