+++ frontend/src/components/Sidepanel/RoomPeople/RoomPersonCard/index.tsx
@@ -2,15 +2,15 @@ import { ChevronDown, Pencil, Trash2, User } from "lucide-react";
 import { AnimatePresence, motion } from "motion/react";
 import type { CSSProperties } from "react";
 import { useEffect, useId, useReducer } from "react";
-import { useRoomSelection } from "../hooks/useRoomSelection";
-import type { RoomContract } from "../types";
+import { useRoomSelection } from "../../../../hooks/useRoomSelection";
+import type { RoomContract } from "../../../../types";
 import {
   formatContractDate,
   formatContractStatus,
   getContractDateMeta,
-} from "../utils/contractDates";
-import { EXPAND_COLLAPSE_TRANSITION } from "../utils/motionTransitions";
-import { renderValue } from "../utils/renderValue";
+} from "../../../../utils/contractDates";
+import { EXPAND_COLLAPSE_TRANSITION } from "../../../../utils/motionTransitions";
+import { renderValue } from "../../../../utils/renderValue";
 
 let seenExpandReqId: number | null = null;
