+++ frontend/src/components/Sidepanel/index.tsx
@@ -1,7 +1,7 @@
 import { motion } from "motion/react";
-import { EXPAND_COLLAPSE_TRANSITION } from "../utils/motionTransitions";
+import { EXPAND_COLLAPSE_TRANSITION } from "../../utils/motionTransitions";
 import RoomInfo from "./RoomInfo";
-import RoomPeople from "./RoomPeople";
+import RoomPeople from "./RoomPeople/RoomPeople";
 import "./SidePanel.css";
 
 type SidePanelProps = {
