+++ frontend/src/components/Sidepanel/style.css
